+++ packages/aws-client/src/SQS.ts
@@ -1,7 +1,7 @@
 /**
  * @module SQS
  */
-
+import Logger from '@cumulus/logger';
 import get from 'lodash/get';
 import isObject from 'lodash/isObject';
 import isString from 'lodash/isString';
@@ -12,6 +12,7 @@ import { sqs } from './services';
 import { inTestMode } from './test-utils';
 import { improveStackTrace } from './utils';
 
+const log = new Logger({ sender: '@cumulus/aws-client/SQS' });
 export interface SQSMessage extends AWS.SQS.Message {
   ReceiptHandle: string
 }
@@ -39,7 +40,9 @@ export const getQueueUrlByName = async (queueName: string) => {
 export async function createQueue(QueueName: string) {
   const createQueueResponse = await sqs().createQueue({
     QueueName,
-  }).promise();
+  })
+    .on('error', (error) => log.error(error))
+    .promise();
 
   if (inTestMode()) {
     if (createQueueResponse.QueueUrl === undefined) {
@@ -68,7 +71,9 @@ export async function createQueue(QueueName: string) {
 export const deleteQueue = (queueUrl: string) =>
   sqs().deleteQueue({
     QueueUrl: queueUrl,
-  }).promise();
+  })
+    .on('error', (error) => log.error(error))
+    .promise();
 
 export const getQueueAttributes = async (queueName: string) => {
   const queueUrl = await getQueueUrlByName(queueName);
@@ -94,19 +99,25 @@ export const getQueueAttributes = async (queueName: string) => {
  * @param {string} queueUrl - url of the SQS queue
  * @param {string|Object} message - either string or object message. If an
  *   object it will be serialized into a JSON string.
+ * @param {Logger} [logOverride] - optional Logger passed in for testing
  * @returns {Promise} resolves when the messsage has been sent
- **/
-export const sendSQSMessage = (queueUrl: string, message: string | object) => {
-  let messageBody;
-  if (isString(message)) messageBody = message;
-  else if (isObject(message)) messageBody = JSON.stringify(message);
-  else throw new Error('body type is not accepted');
 
-  return sqs().sendMessage({
-    MessageBody: messageBody,
-    QueueUrl: queueUrl,
-  }).promise();
-};
+ **/
+export const sendSQSMessage = improveStackTrace(
+  (queueUrl: string, message: string | object, logOverride: Logger) => {
+    const logger = logOverride || log;
+    let messageBody;
+    if (isString(message)) messageBody = message;
+    else if (isObject(message)) messageBody = JSON.stringify(message);
+    else throw new Error('body type is not accepted');
+
+    return sqs().sendMessage({
+      MessageBody: messageBody,
+      QueueUrl: queueUrl,
+    }).on('error', (error) => logger.error(error))
+      .promise();
+  }
+);
 
 type receiveSQSMessagesOptions = {
   numOfMessages?: number,
@@ -139,7 +150,8 @@ export const receiveSQSMessages = async (
     MaxNumberOfMessages: options.numOfMessages || 1,
   };
 
-  const messages = await sqs().receiveMessage(params).promise();
+  const messages = await sqs().receiveMessage(params)
+    .on('error', (error) => log.error(error)).promise();
 
   return <SQSMessage[]>(messages.Messages ?? []);
 };
@@ -156,7 +168,9 @@ export const parseSQSMessageBody = (message: any): unknown =>
  */
 export const deleteSQSMessage = improveStackTrace(
   (QueueUrl: string, ReceiptHandle: string) =>
-    sqs().deleteMessage({ QueueUrl, ReceiptHandle }).promise()
+    sqs().deleteMessage({ QueueUrl, ReceiptHandle })
+      .on('error', (error) => log.error(error))
+      .promise()
 );
 
 /**
@@ -177,7 +191,11 @@ export const sqsQueueExists = async (queueUrl: string) => {
     await sqs().getQueueUrl({ QueueName }).promise();
     return true;
   } catch (error) {
-    if (error.code === 'AWS.SimpleQueueService.NonExistentQueue') return false;
+    if (error.code === 'AWS.SimpleQueueService.NonExistentQueue') {
+      log.warn(`Queue ${QueueName} does not exist`);
+      return false;
+    }
+    log.error(error);
     throw error;
   }
 };